# app/Http/Controllers/Api/v1/BuyBackController.php
<?php

namespace App\Http\Controllers\Api\v1;

use Illuminate\Http\Request;
use App\Http\Controllers\Controller;
use Illuminate\Support\Facades\Validator;
use App\Customer;
use App\Order;
use App\OrderProduct;
use App\ReturnExchangeProduct;
use App\ReturnExchange;
use App\StoreWebsiteOrder;
use Illuminate\Validation\Rule;


class BuyBackController extends Controller
{
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function index()
    {
        //
    }

    /**
     * Show the form for creating a new resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function create()
    {
        //
    }

    /**
     * @SWG\Post(
     *   path="/return-exchange-buyback/create",
     *   tags={"Orders"},
     *   summary="Create return exchange buyback",
     *   operationId="create-return-exchange-buy-back",
     *   @SWG\Response(response=200, description="successful operation"),
     *   @SWG\Response(response=406, description="not acceptable"),
     *   @SWG\Response(response=500, description="internal server error"),
     *      @SWG\Parameter(
     *          name="mytest",
     *          in="path",
     *          required=true, 
     *          type="string" 
     *      ),
     * )
     *
     */
    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     */
    public function store(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'order_id' => 'required|exists:order_products,order_id',
            'product_sku' => 'required|exists:order_products,sku',
            'website' => 'required',
            'type' => 'required|in:refund,exchange,buyback,return'
        ]);

        if ($validator->fails()) {
            return response()->json(['status' => 'failed', 'message' => 'Please check validation errors !', 'errors' => $validator->errors()], 400);
        }

        $storeWebsite = \App\StoreWebsite::where("website","like",$request->website)->first();
        if($storeWebsite) {
            $getCustomerOrderData = StoreWebsiteOrder::Where('platform_order_id', $request->order_id)
                ->where('op.sku', $request->product_sku)->where('store_website_orders.website_id', $storeWebsite->id)
                ->join('orders as od', 'od.id', 'store_website_orders.order_id')
                ->join('order_products as op', 'op.order_id', 'od.id')
                ->join('products as p', 'p.id', 'op.product_id')
                ->select('p.name as product_name', 'op.product_price', 'op.sku', 'op.order_id','op.id as order_product_id', 'op.product_id','od.customer_id')
                ->first();

            if (!isset($getCustomerOrderData) || empty($getCustomerOrderData)) {
                return response()->json(['status' => 'failed', 'message' => 'No order found for the customer!'], 404);
            }
            
            $return_exchange_products_data = [
                "status_id" => 1, //Return request received from customer 
                "product_id" => $getCustomerOrderData->product_id,
                "order_product_id" => $getCustomerOrderData->order_product_id,
                "name" => $getCustomerOrderData->product_name
            ];
            $return_exchanges_data = [
                'customer_id' => $getCustomerOrderData->customer_id,
                'type' => $request->type,
                'reason_for_refund' => $request->get('reason',''.ucwords($request->type).' of product from '.$storeWebsite->website),
                'refund_amount' => $getCustomerOrderData->product_price,
                'status' => 1,
                'date_of_request' => date('Y-m-d H:i:s')
            ];

            $success = ReturnExchange::create($return_exchanges_data);
            if (!$success) {
                return response()->json(['status' => 'failed', 'message' => 'Unable to create '.ucwords($request->type).' request!'], 500);
            }
            ReturnExchangeProduct::create($return_exchange_products_data);

            // send emails 
            try {
                if($request->type == "refund") {
                    \MultiMail::to($success->customer->email)->send(new \App\Mails\Manual\InitializeRefundRequest($success));
                    $view = (new \App\Mails\Manual\InitializeRefundRequest($success))->build();
                    $params = [
                        'model_id'          => $success->id,
                        'model_type'        => \App\ReturnExchange::class,
                        'from'              => $view->fromMailer,
                        'to'                => $success->customer->email,
                        'subject'           => $view->subject,
                        'message'           => $view->render(),
                        'template'          => 'refund-request',
                        'additional_data'   => $success->id
                    ];
                    \App\Email::create($params);
                    \App\CommunicationHistory::create([
                        'model_id'      => $success->id,
                        'model_type'    => \App\ReturnExchange::class,
                        'type'          => 'refund-request',
                        'method'        => 'email'
                    ]);
                }else if ($request->type == "return") {
                    \MultiMail::to($success->customer->email)->send(new \App\Mails\Manual\InitializeReturnRequest($success));
                    $view = (new \App\Mails\Manual\InitializeReturnRequest($success))->build();
                    $params = [
                        'model_id'          => $success->id,
                        'model_type'        => \App\ReturnExchange::class,
                        'from'              => $view->fromMailer,
                        'to'                => $success->customer->email,
                        'subject'           => $view->subject,
                        'message'           => $view->render(),
                        'template'          => 'return-request',
                        'additional_data'   => $success->id
                    ];
                    \App\Email::create($params);
                    \App\CommunicationHistory::create([
                        'model_id'      => $success->id,
                        'model_type'    => \App\ReturnExchange::class,
                        'type'          => 'return-request',
                        'method'        => 'email'
                    ]);
                }else if ($request->type == "exchange") {
                    \MultiMail::to($success->customer->email)->send(new \App\Mails\Manual\InitializeExchangeRequest($success));
                    $view = (new \App\Mails\Manual\InitializeExchangeRequest($success))->build();
                    $params = [
                        'model_id'          => $success->id,
                        'model_type'        => \App\ReturnExchange::class,
                        'from'              => $view->fromMailer,
                        'to'                => $success->customer->email,
                        'subject'           => $view->subject,
                        'message'           => $view->render(),
                        'template'          => 'exchange-request',
                        'additional_data'   => $success->id
                    ];
                    \App\Email::create($params);
                    \App\CommunicationHistory::create([
                        'model_id'      => $success->id,
                        'model_type'    => \App\ReturnExchange::class,
                        'type'          => 'exchange-request',
                        'method'        => 'email'
                    ]);
                }

            }catch(\Exception $e) {
                \Log::channel('customer')->info("Sending mail issue at the buybackcontroller #158 ->".$e->getMessage());
            }
            return response()->json(['status' => 'success', 'message' => ucwords($request->type).' request created successfully'], 200);
        }else{
            return response()->json(['status' => 'failed', 'message' => 'Please check website is not exist'], 404);
        }
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function show($id)
    {
        //
    }

    /**
     * Show the form for editing the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function edit($id)
    {
        //
    }

    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function update(Request $request, $id)
    {
        //
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function destroy($id)
    {
        //
    }

    /**
     * @SWG\Get(
     *   path="/orders/products",
     *   tags={"Orders"},
     *   summary="Check product for buyback",
     *   operationId="check-product-for-buy-back",
     *   @SWG\Response(response=200, description="successful operation"),
     *   @SWG\Response(response=406, description="not acceptable"),
     *   @SWG\Response(response=500, description="internal server error"),
     *      @SWG\Parameter(
     *          name="mytest",
     *          in="path",
     *          required=true, 
     *          type="string" 
     *      ),
     * )
     *
     */
    
    public function checkProductsForBuyback(request $request)
    {
        $validator = Validator::make($request->all(), [
            'customer_email' => 'required|email',
            'website' => 'required'
        ]);
        if ($validator->fails()) {
            return response()->json(['status' => 'failed', 'message' => 'Please check validation errors !', 'errors' => $validator->errors()], 400);
        }

        $responseData = [];

        $storeWebsite = \App\StoreWebsite::where("website","like",$request->website)->first();
        if($storeWebsite) {
            $checkCustomer = Customer::where('email', $request->customer_email)
            ->where('store_website_id',$storeWebsite->id)
            ->first();

            if (!$checkCustomer) {
                return response()->json(['status' => 'failed', 'message' => 'Customer not found with this email !'], 404);
            }

            $customer_id = $checkCustomer->id;
            $getCustomerOrderData = Order::Where('customer_id', $customer_id)->where('swo.website_id',$storeWebsite->id)
                ->join('order_products as op', 'op.order_id', 'orders.id')
                ->join('products as p', 'p.id', 'op.product_id')
                ->join('store_website_orders as swo', 'swo.order_id', 'op.order_id');
                
            if($request->order_id !=  null) {
               $getCustomerOrderData = $getCustomerOrderData->where("swo.platform_order_id",$request->order_id);
            }

            $getCustomerOrderData = $getCustomerOrderData->select('p.name as product_name', 'op.product_price', 'op.sku',"op.id as order_product_id", 'op.product_id','swo.platform_order_id as order_id')
                ->get()->makeHidden(['action']);

            if (count($getCustomerOrderData) == 0) {
                return response()->json(['status' => 'failed', 'message' => 'No order found for the customer!'], 404);
            }
            $responseData = [];
            foreach($getCustomerOrderData as $getCustomerOrder){
                $responseData[$getCustomerOrder->order_id][] = $getCustomerOrder;
            }
        }


        return response()->json(['status' => 'success', 'orders' => $responseData], 200);
    }
}