# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Live Test",
        "description": "This is api testing",
        "contact": {
            "email": "xyz@xyz.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/affiliate/add": {
            "post": {
                "tags": [
                    "Affiliate"
                ],
                "summary": "store affiliate",
                "operationId": "store-affiliate",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/influencer/add": {
            "post": {
                "tags": [
                    "Influencer"
                ],
                "summary": "store influencer",
                "operationId": "store-influencer",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/return-exchange-buyback/create": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create return exchange buyback",
                "operationId": "create-return-exchange-buy-back",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/orders/products": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Check product for buyback",
                "operationId": "check-product-for-buy-back",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}